Assegnazione Entity a OggettiInventario e OggettiBattleGround

+++ src/main/java/com/adryanSoftwares/ilGiocoDelTizioCheStaConAdryan/Dto/ChestDTO.java
@@ -16,7 +16,7 @@ import java.util.List;
 @NoArgsConstructor
 public class ChestDTO {
     @JsonProperty("idChest")
-    private Long idChest;
+    private Integer idChest;
     @JsonProperty("dimensione")
     private String dimensione;
     @JsonProperty("livello")
